fixing setup.py failure caused by uninstalled dependencies in constants.py
