+++ .githooks/_pre_commit/_censor_creds.py
@@ -1,7 +1,8 @@
 import json
 import re
 # local modules
-from all_repos_add_readme.constants import GithubConstants, GitConstants
+from all_repos_add_readme.constants import GithubConstants
+from all_repos_add_readme.github_utils.git_utils import GitConstants
 from all_repos_add_readme.github_utils.git_utils import git_config
