+++ all_repos_add_readme/main.py
@@ -8,7 +8,7 @@ import logging
 from all_repos_add_readme.github_utils import github_api
 from all_repos_add_readme._exceptions import InvalidReadme, ExceptionMessages
 from all_repos_add_readme.constants import Constants
-from all_repos_add_readme.constants import LoggerConstants
+from all_repos_add_readme.logger_constants import LoggerConstants
 from all_repos_add_readme.constants import ToolArgumentNames
 from all_repos_add_readme._logger import setup_logger
 from all_repos_add_readme._logger import shutdown_logging
